chore: Add ChakraProvider wrapper component
This commit adds a new file `providers.tsx` which contains the `Providers` component. This component wraps the entire application with the `ChakraProvider` from `@chakra-ui/react`. This is necessary to provide the Chakra UI theme and styling to all components in the application.

Dependencies have also been updated in the `package.json` file to include the required packages for Chakra UI.

The purpose of this commit is to set up the necessary infrastructure for using Chakra UI in the application.

```